SWPWA-561 - Desktop menu redesign
