+++ src/app/component/Menu/index.js
@@ -9,4 +9,4 @@
  * @link https://github.com/scandipwa/base-theme
  */
 
-export { default } from './MenuOverlay.container';
+export { default } from './Menu.container';
